Add restreamer endpoint for viewing chat messages
